+++ src/linux/wayland.rs
@@ -1,10 +1,12 @@
-use std::collections::VecDeque;
-use std::convert::TryInto;
-use std::env;
-use std::os::unix::io::AsFd;
-use std::os::unix::net::UnixStream;
-use std::path::PathBuf;
-use std::time::Instant;
+use std::{
+    collections::VecDeque,
+    convert::TryInto as _,
+    env,
+    num::Wrapping,
+    os::unix::{io::AsFd, net::UnixStream},
+    path::PathBuf,
+    time::Instant,
+};
 
 use log::{debug, error, trace, warn};
 use wayland_client::{
@@ -32,7 +34,7 @@ pub struct Con {
     event_queue: EventQueue<WaylandState>,
     state: WaylandState,
     virtual_keyboard: Option<zwp_virtual_keyboard_v1::ZwpVirtualKeyboardV1>,
-    input_method: Option<(zwp_input_method_v2::ZwpInputMethodV2, u32)>,
+    input_method: Option<zwp_input_method_v2::ZwpInputMethodV2>,
     virtual_pointer: Option<zwlr_virtual_pointer_v1::ZwlrVirtualPointerV1>,
     base_time: std::time::Instant,
 }
@@ -141,7 +143,7 @@ impl Con {
                 .state
                 .im_manager
                 .as_ref()
-                .map(|im_mgr| (im_mgr.get_input_method(seat, &qh, ()), 0));
+                .map(|im_mgr| im_mgr.get_input_method(seat, &qh, ()));
         };
 
         // Setup virtual pointer
@@ -306,7 +308,7 @@ impl Drop for Con {
         if let Some(vk) = self.virtual_keyboard.take() {
             vk.destroy();
         }
-        if let Some((im, _)) = self.input_method.take() {
+        if let Some(im) = self.input_method.take() {
             im.destroy();
         }
         if let Some(vp) = self.virtual_pointer.take() {
@@ -327,6 +329,7 @@ impl Drop for Con {
 struct WaylandState {
     keyboard_manager: Option<zwp_virtual_keyboard_manager_v1::ZwpVirtualKeyboardManagerV1>,
     im_manager: Option<zwp_input_method_manager_v2::ZwpInputMethodManagerV2>,
+    im_serial: Wrapping<u32>,
     pointer_manager: Option<zwlr_virtual_pointer_manager_v1::ZwlrVirtualPointerManagerV1>,
     seat: Option<wl_seat::WlSeat>,
     /*  output: Option<wl_output::WlOutput>,
@@ -438,13 +441,17 @@ impl Dispatch<zwp_input_method_manager_v2::ZwpInputMethodManagerV2, ()> for Wayl
 }
 impl Dispatch<zwp_input_method_v2::ZwpInputMethodV2, ()> for WaylandState {
     fn event(
-        _state: &mut Self,
+        state: &mut Self,
         _vk: &zwp_input_method_v2::ZwpInputMethodV2,
         event: zwp_input_method_v2::Event,
         (): &(),
         _: &Connection,
         _qh: &QueueHandle<Self>,
     ) {
+        match event {
+            zwp_input_method_v2::Event::Done => state.im_serial += Wrapping(1u32),
+            _ => (), // TODO
+        }
         warn!("Got a input method event {:?}", event);
     }
 }
@@ -540,15 +547,18 @@ impl Drop for WaylandState {
 
 impl Keyboard for Con {
     fn fast_text(&mut self, text: &str) -> InputResult<Option<()>> {
-        let Some((im, serial)) = self.input_method.as_mut() else {
+        let Some(im) = self.input_method.as_mut() else {
             return Ok(None);
         };
 
         is_alive(im)?;
         trace!("fast text input with imput_method protocol");
+        // Process all previous events so that the serial number is correct
+        self.event_queue
+            .roundtrip(&mut self.state)
+            .map_err(|_| InputError::Simulate("The roundtrip on Wayland failed"))?;
         im.commit_string(text.to_string());
-        im.commit(*serial);
-        *serial = serial.wrapping_add(1);
+        im.commit(self.state.im_serial.0);
 
         // TODO: Change to flush()
         self.event_queue